requirements.txt
python-apt==2.4.0+ubuntu4
python-crontab==3.2.0
python-dotenv==1.0.1
python-requests==2.31.0
python-urllib3==2.0.4
python-psutil==6.0.0
python-pycurl==8.1.2
